requirements.txt
click>=7.1.2
python-dotenv>=1.0.0
os-sys